Restore wpe-2.22 visibility change behavior
The change of IsInWindow activity state  could suspend active DOM
objects and so make hide/show API to behave similar to
suspend/resume. The web application is not notified about visibility
state in this case, which is regression compared to wpe-2.22.

This change tries to restore previous behavior. The IsVisible activity
state is used to notify about visibility changes and also suspend
painting.