fix video freeze after initial seek
webKitMediaSrcPrepareInitialSeek can be called after appsrc already
invoked initial 'seek-data' and 'need-data' callbacks. this leaves
webkitmediasrc in inconsitent state, with appsrcSeekDataCount = 0 and
appsrcSeekDataNextAction = MediaSourceSeekToTime.

Because appsrcSeekDataNextAction is set, it blocks ReadyForMoreSamples
notifications. This leads to starvation of video source and
video freeze (after the first 'enough-data').

This fix removes setup of appsrcSeekDataNextAction and
appsrcNeedDataFlag. These are not really required in this case, beacuse:
 - initial seek doesn't block source buffers
 - MediaPlayerPrivateGStreamerMSE calls notifySeekNeedsDataForTime()
 immediately webKitMediaSrcPrepareInitialSeek()

%% original patch: 0002-MSE-GStreamer-fix-video-freeze-after-initial-seek.patch