MSE: limit the number of samples stored in SourceBuffer
Applications may append a lot of samples without triggereing
eviction algorithm (> 500K). Which results in cleanup(on player
destruction) taking too long time.

Here is an example of video where we observed such behavior:
https://www.youtube.com/tv#/watch?v=6ATfcsAFFNE&list=RD-WLgxoa4BeU

This change allows at least one pass of eviction algorithm if number
of samples is greater than 10K. Which should be enough to allow
storing ~2-3 minutes of audio or video if size limit doesn't kick in.